+++ Site/webapp/wdkCustomization/js/client/components/Record/OntologyRecord/OntologyRecordHeading.scss
