--- Site/webapp/wdkCustomization/js/client/components/Record/Page/RecordHeading/Shared/shared.scss
@@ -1,1 +1,0 @@
-@import './ImpactIndicator/ImpactIndicator';
