--- Site/webapp/wdkCustomization/js/client/components/Record/Page/RecordPageTypes.ts
@@ -1,9 +1,0 @@
-export interface HeaderActions {
-    iconClassName: string;
-    onClick: any;
-    label: string;
-}
-
-export const isVariantRecord = (item: any) => {
-    return item.recordClassName === "VariantRecordClasses.VariantRecordClass"
-};
