+++ Site/webapp/wdkCustomization/js/client/components/Record/RecordHeading/ImpactIndicator/ImpactIndicator.scss
