+++ Site/webapp/wdkCustomization/js/client/components/Record/RecordHeading/ImpactIndicator/ImpactIndicator.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 
+import "./ImpactIndicator.scss";
+
 interface ImpactIndicator {
 	impact: string;
 }
 
-const ImpactIndicator: React.SFC<ImpactIndicator> = props => {
+export const ImpactIndicator: React.SFC<ImpactIndicator> = props => {
 	const { impact } = props;
 	return <span className={`impact-indicator ${impact.toLowerCase()}`}>{impact}</span>
 }
 
-export default ImpactIndicator;
