+++ Site/webapp/wdkCustomization/js/client/components/Record/RecordTable/CssBarChart/CssBarChart.tsx
