+++ Site/webapp/wdkCustomization/js/client/components/Record/RecordTable/RecordTable.tsx
