+++ Site/webapp/wdkCustomization/js/client/components/Record/Sections/MainCategorySection.tsx
