+++ Site/webapp/wdkCustomization/js/client/components/Record/Sections/RecordNavigationSection.tsx
