+++ Site/webapp/wdkCustomization/js/client/components/Record/Sections/RecordTableSection.tsx
