+++ Site/webapp/wdkCustomization/js/client/components/Record/TrackRecord/TrackRecordHeading.scss
