+++ Site/webapp/wdkCustomization/js/client/components/Record/Types/DatasetRecordClasses.DatasetRecordClass.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import RecordHeading  from '../Page/RecordHeading/DatasetRecord/DatasetRecordHeading';
-import RecordMainSection from '../Page/Sections/RecordMainSection';
-import RecordUI from '../Page/RecordUI';
+import RecordHeading  from '../DatasetRecord/DatasetRecordHeading';
+import RecordMainSection from '../Sections/RecordMainSection';
+import RecordUI from '../RecordUI';
 
 export { RecordHeading, RecordMainSection, RecordUI };
