+++ Site/webapp/wdkCustomization/js/client/components/Record/Types/GeneRecordClasses.GeneRecordClass.tsx
@@ -1,7 +1,7 @@
 
-import RecordHeading from '../Page/RecordHeading/GeneRecord/Heading';
-import RecordMainSection from '../Page/Sections/RecordMainSection';
-import RecordUI from "../Page/RecordUI";
+import RecordHeading from '../GeneRecord/GeneRecordHeading';
+import RecordMainSection from '../Sections/RecordMainSection';
+import RecordUI from "../RecordUI";
 
 
 export { RecordHeading, RecordMainSection, RecordUI };
