+++ Site/webapp/wdkCustomization/js/client/components/Record/Types/OntologyRecordClasses.OntologyRecordClass.tsx
@@ -1,7 +1,7 @@
 
-import RecordHeading from '../Page/RecordHeading/OntologyRecord/OntologyRecordHeading';
-import RecordMainSection from '../Page/Sections/RecordMainSection';
-import RecordUI from "../Page/RecordUI";
+import RecordHeading from '../OntologyRecord/OntologyRecordHeading';
+import RecordMainSection from '../Sections/RecordMainSection';
+import RecordUI from "../RecordUI";
 
 
 export { RecordHeading, RecordMainSection, RecordUI };
