+++ Site/webapp/wdkCustomization/js/client/components/Record/Types/VariantRecordClasses.VariantRecordClass.tsx
@@ -1,5 +1,5 @@
-import RecordHeading from '../Page/RecordHeading/VariantRecord/VariantRecordHeading';
-import RecordMainSection from '../Page/Sections/RecordMainSection';
-import RecordUI from "../Page/RecordUI";
+import RecordHeading from '../VariantRecord/VariantRecordHeading';
+import RecordMainSection from '../Sections/RecordMainSection';
+import RecordUI from "../RecordUI";
 
 export { RecordHeading, RecordMainSection, RecordUI };
