+++ Site/webapp/wdkCustomization/js/client/components/Record/Types/index.ts
@@ -15,6 +15,8 @@ export interface RecordHeading {
 
 export type RecordAttributes = Record<string, AttributeValue>;
 
-/* export const isVariantRecord = (item: any) => {
+
+export const isVariantRecord = (item: any) => {
     return item.recordClassName === "VariantRecordClasses.VariantRecordClass"
-}; */
+};
+
